Minimal fake factset_financial_data dataset that allow overwriting values

Minimal fake factset_financial_data dataset that allow overwriting values

Usage

fake_factset_financial_data(
  fsym_id = NULL,
  isin = NULL,
  factset_entity_id = NULL,
  adj_price = NULL,
  adj_shares_outstanding = NULL,
  issue_type = NULL,
  one_adr_eq = NULL,
  ...
)

Arguments

fsym_id

A FactSet FSYM ID (character).

isin

An ISIN (character).

factset_entity_id

A FactSet Entity ID (character).

adj_price

Adjusted price (numeric).

adj_shares_outstanding

Adjusted shares outstanding (numeric).

issue_type

Issue type (character).

one_adr_eq

One ADR equivalent (numeric).

...

Arguments passed on to tibble::tibble(). These are useful in case you want to add additional columns to the output tibble.

Value

A tibble with the specified columns.

Fake data

These funtions are developer-oriented. They all call tibble() so you can expect all the goodies that come with that.

fake_foo fakes a dataset with the name foo

Params

The arguments are the column names of the datasets being faked. They all have a default and it can be overwritten.

Pros and cons

By default, these functions output a 1-row data-frame, which can be useful for CI/CD. They also help you to avoid duplicating test code, and help the reader of your code to focus on the one thing you want to test, instead of burring that thing in the much longer code you need to create a fake object from scratch. Since the default parameters are abstracted into utils.R, it makes it easier to maintain coordination of values across a variety of datasets.

But fake_*() functions hide the explicit content. If the reader of your code wants to inspect the data being tested, they need to jump to the function definition or call them interactively.

Examples

fake_factset_financial_data()
#> # A tibble: 1 × 7
#>   fsym_id  isin    factset_entity_id adj_price adj_shares_outstanding issue_type
#>   <chr>    <chr>   <chr>                 <dbl>                  <dbl> <chr>     
#> 1 A12345-B AB123C… 1234A5-B               12.3             1234567890 EQ        
#> # ℹ 1 more variable: one_adr_eq <dbl>

fake_factset_financial_data(isin = "US0123456789")
#> # A tibble: 1 × 7
#>   fsym_id  isin    factset_entity_id adj_price adj_shares_outstanding issue_type
#>   <chr>    <chr>   <chr>                 <dbl>                  <dbl> <chr>     
#> 1 A12345-B US0123… 1234A5-B               12.3             1234567890 EQ        
#> # ℹ 1 more variable: one_adr_eq <dbl>